History to Get the Own Base Elasticity Plot
I conducted this a few weeks ago.  But, essentially, what this R history was able to do was create an Own Tax Base Elasticity w.r.t the Sales Factor Apportionment weight.  I filtered and mutated the apportionment data to create percentage change in both revenue and the sales factor.  Then, calculated an elasticity for each change.  Then, I plotted that elasticity for each state.  This did not control for State, time, or rate.   I am going to try that next.