Clean Weights and Dates for JMP
As of this date, this is the cleanest version of the data for the Sales Factor weights and effective dates for the policy.

The Sales Factor weights are the effective statutory formula weights in place for each state.  I include dummy variables for years that states allowed (a) different industries to use another formula, (b) election of either formula, and (c) phase-ins that impacted usage of the statutory formula.  The phase in is controlled for by the weight, but this is if the use of the formula is somehow tied to collections or some other differentiating factor.

The Date pages are for single sales factor implementation and DWS.  The single sales factor dates are correct and complete, to the best of my knowledge.  The date page includes columns for phase-in years and election.   The date page for DWS is not complete.  The research required to find the effective dates going back that far is substantial, and right now is preventing me from moving forward with a result.

Other than any possible updates, I think this will be the final outside data I use for the JMP.