History from session on non-log result
I attempted to a get a simple non log revenue result.  However, I was unsuccessful.  That is because there is not enough variation in the Sales factor to distinguish it from the state result.  I am working on it.  But, I need to research the log linear regression more to see if that will solve or just prolong the problem.