Save Data for the Elasticity Attempt 1
This is the data and R file that I ran to try and find the elasticity coefficient.  Pausing on finding the elasticity this rout.  I am going to run the simple result that dougan asked for.  Estimate the impact of these changes on Log (Total Revenue).   I may use a dummy variable or a simple DiD.