History from Prelim Result
This is the history from running the preliminary result.  I attempted Callaway, 2 way FE, and just settled with a simple OLS-FE regression.  I needed the result.  I will conduct robustness checks on it, and plots with the data.  It's possible there is something here.  The sales factor was significant, but the post on the policy dummy was not (likely drowned out by noise from year/early adopters).