Log Total Rev Coefficient Estimate History
There is a lot to summarize in this save:

1) The most important results from this session are:

a) revenue_panel, which is Res3.  It is the foundation that yielded the estimates for the impacts of the policy.

b) state_post_dummy_total_rev. Is a cleaned dataframe that estimates the effect of the policy for each state.

Relative to my previous results that I sent to Dougan, those results set Alabama as the base state case.  These results disaggregate interaction relative to Alabama.  Therefore, this set of results is more complete.  I have entered them into Latex.

2). I also saved these dataframes a second time wtih the date.

3). The Sales_factor 3WFE includes cuts from the main R script for dougan.  he just wanted a policy dummy variable.  So I included the regressions that left the sales factor as a numeric element of the regression in this 3wFE sheet.

Next step, I will load the Res3 and coefficient dataframes.  From there, I will create a chart that sorts them to see if a pattern emerges.  I will need to merge the regions and possibly industries onto those states.  I will start with regions for Dougan.  Then I will send a basic ggplot that organizes that information.

Then onto an attempt to create the canonical DiD plot.  I know how I will will create the plot for the treated states.  However I am not not sure how I will come up with the changes for the untreated group.  stay tuned.