Severance Tax Results continued
I got close to estimating the severance tax result.  I was able to estimate the effect.  However, the states that do not have treatment are filtered out in the post_coefficients dataframe. I will need to restart with just a simple regression, which is what I think Dougan is going for.