Bootstrap Loop
I was unable to figure out the bootstrap component of the loop for standard errors.  I can compute it individually.  But, for now, I am going with the standard error created from the sample estimate.  So, functionally, assuming normality.  I can mess around with the bootstrapping on the flight.

For now, going to compile results for these estimates and try to get results for share of Collections tonight.