Updates to History
Just re-ran the original loop that I had before Germany, and it ran.  No problems.  No need to add the one on the dataframe for the first dataframe filtering part of the loop.

So, maybe it is the Revenue dataframe.  I will see if I can get the result from the CORP dataframe.

Although, that is interesting because for the shares that did run on the plane, it was from the filtered REV dataframe.  I will commit this, and run the Plane loop and see what happens.