Update naive_CI_sDiD.R
There is an issue computing the Naive CI annual summation.  For years 2009 to 2013, the summation is giving the result "Inf".  This is preventing me from running the sDiD for any adopters after that period.  It came up in the summary stats as zero.

I will work on this more in the morning.