Results for 9/15 Draft
The results I obtained for the 9/15 draft are the following:
1) Real, Naive CI per capita, the associated plots, time and date weights.
2) Percentage Change in Real, Naive CI, the associated significant plots, time and date weights.
Note: I think this is not the best way to estimate this change.  I think I should have a base year and estimate percentage change from that base year.  Further, I think I should just use a TWFE with 3 years of dummies after for a treated state (then possibly, somehow, eliminating that state, not sure how it would work yet).

3) Nationwide Share of Real, Naive CI, associated significant plots, and  time and state weights.