Ilinois Trial- syn vs Actual
The trial was fine.  However, clearly, over time, more corporate income is moving to Florida.  This result gives credence to weighting the real income by per capita.  Or else, I won't be able to control for the fact that Florida just gets more Corporate Income irrespective of it's apportionment policy.

The per capita is a way of moderating this influence.