rout forward= real CI per capita to create Controls
I will use this estimation for Illinois as the basis to compute the percentage change.

I will use the point estimate from the 3 year to compute the percentage change in revenue in the short run.

Then I will use the controls created prior to the switch to create a long run synthetic Illinois (treatment state).

This will be used to compute the Long percentage change in revenue.

It can clearly be done.  It is not perfect.  It will require for the 28 states, carefully looking at pre-trends (make sure they pass the eye test).  Possibly compute the test discussed by Micah.  Every state that switched after 2017 will not have a long run effect.  Will just compute the 3 year short run window for them.