History sDiD- Attempt to Get Percent Change
Attempted to get percentage change in revenue from sDiD package.

Will need to modify and use a the synthetic control for Percentage Change.  I might revisit this and redo the point estimates for the effect of the policy in the short run, but right now, am going all in on getting the SC to work.