Tried Shift up at Treatment Year
The shift up at treatment year, did not work.  Therefore, I will need to attempt normalizing everything to 1976.  I am not sure whether that should be done as per capita.  Right now,  will do so.

Either way, in this commit, check the plot for the shift at treatment year.  The pre-trends are not good. I will attempt this on the 1976 base year.  If that doesn't work, I guess the ARIMA model is the best way to go or matching or MASC.