Base year code 1976
I do not think this is a perfect way to estimate percentage change in revenue.  In fact, I think the next step is to compute the percentage change in revenue from an ARIMA model or another matching technique.  This is, without question, not a good synthetic control Illinois.  But, it is an "okay" match.