sDID long run
The sDiD in the long run makes a big difference. Even when the control group isn't big enough in terms of revenue to cover the size of Illinois, it gives a negative point estimate in the long run for Illinois, when it gives them a positive point estimate in the short run.

I am also saving the normalized Illinois.