Create pe_result_IndInc_SR_LR.R
This is the script for finding point estimates for individual income tax.

The Rev dataframe does not include the the no income tax states.  Therefore, this will require about a half day of cleaning/sorting in order to get results.

The same will apply for finding the total.  Will try again for the percentage change in Corporate Income.