Update per_chan_CI.R
This update finally got the percentage change in Corporate Income in the SR and LR.

Tomorrow, I will need to finish the loop for the plots.  That will be the final validation.

Second, I will compute a series of percentage change in averages.  I will computes the average SR and LR effects by (a) year cohort of adoption and (b) region.

Once this has been done, I will be able to compute it for Non corporate tax revenue next.

Any jobs or robustness results will have to wait until after the workshop, but this is moving downfield, rapidly.