R Scripts for Percentage Change in CI
These scripts get the percentage change in Corporate Income.  The file for Iowa is the trial and per_chan script includes the 6 loops for cleaning the dataframes, getting the weights, creating the synthetic control, creating the optimal shift, computing percentage change, and plotting each of those.

The dataframes created with this script were saved last night in the Dataframes to plot for percentage change commit