Federal and State Corporate Income
It seems like the year for either Federal or State collections is one year off.

It think it comes down to both of them being from different sources.

state is from Census survey of States.

Federal is from the BEA and is currently the quarterly.

I used the average aggregation in FRED.  Though, doesn't remove the difference.  Might plot the collections summed directly from FRED without any modifications to the date next.
https://fred.stlouisfed.org/series/B075RC1Q027SBEA#0