Stacked DiD Point Estimate & Plot
I was able to get a stacked DiD and point estimate run.  However, naive CI was not inflation adjusted.  It makes me want to go back and redo the (1) descriptive result, (2) the SDID, and now  (3) the Stacked DID.

Right now, I need to just take it one step at at a time.  I will start with the Stacked adjusted for inflation.  I will then detrend it.

(2) SDID

(3) Redo the descriptive.