Notes from Jonathan on Needs for Approach
This is the note from the meeting with Jonathan for the next update to this section:

Dit in TWFE should be an indicator.

Create variable Dst which is an indicator for being a treated state and in the post period and it is a zero otherwise.

For Simple DiD- do not include fixed effects.  One if indicator for treated state. One is an indicator for being post period.  Other is the those interacted.

Just 2006 group.

Just 2007 group.

For event-study specification.  Make sure.  Manually create variables that are indicators for being a treated state in a particular year.

Variable Dst^k is 1 if s is treated and if t = k, 0 otherwise.  So it is an indicator

Create "Treated

Create a regression that treated_1990 and treated_1991 through 2014.  Then exclude the 2005 one for the 2006 analysis.  Exclude the 2006 one for the 2007 analysis.

____

Decrease as an eyeball.

2006 through 2014 compared to the ones in the pre-period.

Simple Regression to quantify the whole thing.  It should be around what you are seeing.

___

Double check the TWFE from the Data Center.